Issue 357 - Coalesced builds should be the same as non-existant builds
In the for loop where we try to evaulate if there is a build for a given
push we don't detect properly that a coalesced build should not count
as an existant build.

This causes us to run test jobs on builds for the push that actually
run the build.

This causes the tests to be run against a wrong build.

This code change ignores coalesced builds.