Allow appending Buildbot children to a TC task
The buildbot_to_taskcluster.py script allows you to append Buildbot
builders to a given TC.

In order to append all downstream builders for a given Buildbot build to
a TC task you would need to list every single one of the builders.

To make this process easier, instead of passing all children builders,
we ask instead for "give me all children for this parent builder".
The option added is --children-of which will determine all children
for a given build.