Bug 1195824 - Deal with permanent failures properly.
Up until now, we would add extra load when automatically backfilling
for permanent failures.
We fill in any pushes without the job without being able to tell that the job
has been failing for a long time.

This change looks further than max_revisions to determine if the failure
is long lived.
In such cases we return an empty list of revisions indicating that there is
nothing to backfill for.