Improve automated backfilling's performance and output
When running automated backfilling, it is hard to go through so much
output.
In this patch, we do a bit of code clean up plus adjusting the logging
levels.

We also add the ability to disable validation of buildernames and
revisions
since we can safely assume that Treeherder can only send valid revisions
and buildernames to pulse_actions.

We also add the ability to disable the progress bar when downloading
files over the network. This is important as papertrail's logging
gets ugly because of it.

We also remove inspecting for a good job to double of revisions of
max_revisions.
This is important as we can slow down the processing and we are likely
wasting our time.