Bug 1203552 - Specify to which task to upload artifacts to
In bug 1221091 we lost the ability for associated TC tasks containing
all properties of a Buildbot job. This is necessary for test jobs
run under Buildbot would work when scheduled through the BBB.

Until bug 1221091 is fixed, we can land some Mozharness changes
that will allow us to specify on the build job a known taskId to
upload artifacts to (including the Buildbot properties file).

This version of Mozci is experimental and will only be used until
bug 1221091 is fixed and we will remove these changes at that point.