Bug 1253700 - Allow using graph metadata in the metadata of tasks.
In a previous release of mozci, we changed how we set the metadata of
the graph versus the metadata of the tasks.

Unfortunately, when generating the metadata of the tasks we grabbed
information from the pushlog and there are certain pushes that are
pushed by the automation. This would not be an issue, except that
TaskCluster requires the owner field in the metadata to be an email
address.

In order to work around this issue for now, we're going to use the
metadata of the graph which we know that on pulse_actions always
includes an email address.

In following releases we should validate the tasks at the mozci level
and add tests for this by using few sample graphs and validate against
the json schema. To be fixed in
https://github.com/mozilla/mozilla_ci_tools/issues/437