Speed up buildbot_bridge.py test + add logging to tests
The original test for buildbot_bridge did not completely mock everything
that it should. That caused the test to take over 10 seconds to run. In
order to speed this up I narrowed down the test to the creation of a
task and the appropriate setting of the metadata for it (original test
case).

We've also added a test module called helpers.py which allows tests to
easily activate adding debug messages. Doing this helps a lot when using
py.test with the -s flag (do not capture stdout). This helps when trying
to determine if functions are properly patched or not.

I've also renamed all tests making reference to a 'repo' repo name and
instead I'm using 'try' to make string matching easier.  I guess I could
have used 'mock-repo'.