Separate the mozci scripts into their own package + add requirements.txt (#455)
I want to separate the mozci scripts into their own package to simplify
the layout of the project plus manage its own dependencies in the
future.

We also add a requirements.txt file to manage the dependencies of mozci.
Include requirements.txt in MANIFEST.in to make it part of the packaging
process.

Remove ignoring .txt files as in the future I would like to remove
anything that creates unnecessary .txt files in the project's root
directory. This will make it more obvious.