When backfilling talos jobs only fills gaps in previous pushes
For backfilling test jobs, we want to find the range of pushes from the
oldest orange job to the last known green job for that test job.

In the case of talos, we lack jobs in between pushes but we don't care
about the status of any of the jobs. We just want to complete the lack
of coverage.

With this change we don't filter pushes for talos jobs.