Give mozci the ability to run `python setup.py test` to run tests (#463)
Up until now, the best way to run tests and work on them was to run tox
once, active its virtualenv and then start making changes to the
tests and run pytest.

With these changes you can now run `python setup.py test` inside of a
virtualenv and it will get all the packages you need plus all the
packages required to run the tests.

In other words, you don't need to run tox to run the tests. However,
we will keep tox since it allows for multiple configurations and it fits
well with the Travis automation jobs.