Fix TaskClusterManager web authentication + other improvements
We were not instantiating TaskCluster's Queue() properly since we had to
call it as {'credentials': credentials}.

The conditions to use the various authentication approaches was also
bogus.

We added web_auth to the class in order that we can control when to use
web authentication.

We've also fixed schedule_task() to return a Task Status Structure
instead of nothing.