Bug 1303071 Improve debugability of triggering talos jobs
We have two talos related functions. One of them is to add all talos
jobs for all builds on a revision while the other one schedules all
talos jobs for one single build.

In this change we add info messages to both to help differentiate them
right away.

We also stop raising MozciError and instead we return a status code.
We also log the exceptions raised when trying to schedule a talos job
which fails.