Do not schedule downstream jobs without upstream jobs
Downstream tasks (Buildbot test and talos builders) that do not have an upstream task
(Buildbot build builder) is because the build that triggers it is executed in the TaskCluster
CI rather than on the Buildbot CI. For instance, Linux64 talos jobs are executed on Buildbot
while the Linux64 builds are generated by TaskCluster. The decision task schedules the
Buildbot downstream jobs via the Buildbot bridge.

In this change we change determine_trigger_objective to ignore try to schedule Buildbot builders
that do not have a Buildbot upstream builder but a TaskCluster upstream task.